requirements.txt
aiocoap==0.4.3
psycopg2-binary